move generic models to rightful location
