+++ src/query_monitor/left_bounded.py
@@ -1,11 +1,62 @@
 """Implementation of BaseQueryMonitor for queries beginning from StartTime"""
+from __future__ import annotations
 import urllib.parse
+from enum import Enum
 
 from duneapi.types import QueryParameter
 
-from src.models import LeftBound
 from src.query_monitor.base import QueryData
 from src.query_monitor.result_threshold import ResultThresholdQuery
+
+
+class TimeUnit(Enum):
+    """
+    Enum representing SQL interval time units
+    """
+
+    SECONDS = "seconds"
+    MINUTES = "minutes"
+    HOURS = "hours"
+    DAYS = "days"
+    WEEKS = "weeks"
+    MONTHS = "months"
+
+    @classmethod
+    def options(cls) -> list[str]:
+        """Returns a list of all available enum items"""
+        return [str(e.value) for e in cls]
+
+
+class LeftBound:
+    """Left Bound for Query Monitor"""
+
+    def __init__(self, units: TimeUnit, offset: int):
+        self.units = units
+        self.offset = offset
+
+    def as_query_parameters(self) -> list[QueryParameter]:
+        """Returns DuneQueryParameters for object instance"""
+        return [
+            QueryParameter.enum_type(
+                name="TimeUnits",
+                value=str(self.units.value),
+                options=TimeUnit.options(),
+            ),
+            QueryParameter.number_type(name="Offset", value=self.offset),
+        ]
+
+    @classmethod
+    def from_cfg(cls, cfg: dict[str, int]) -> LeftBound:
+        """Loads LeftBound based on dict containing keys units and offset"""
+        return cls(
+            units=TimeUnit(cfg["units"]),
+            offset=int(cfg["offset"]),
+        )
+
+    def __eq__(self, other: object) -> bool:
+        if isinstance(other, LeftBound):
+            return self.units == other.units and self.offset == other.offset
+        raise ValueError(f"Can't compare LeftBound with {type(other)}")
 
 
 class LeftBoundedQueryMonitor(ResultThresholdQuery):
